# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Bulk Purchasing Platform API
  description: |
    Complete API documentation for the Save Together bulk purchasing application.
    
    ## Features
    - User authentication with JWT tokens
    - Product catalog management
    - Vendor management and dashboards
    - Bulk purchasing groups
    - Order management with MOQ validation
    - Escrow-protected transactions
    - Dispute resolution
    - Real-time chat with WebSocket (Socket.IO)
    
    ## Authentication
    Most endpoints require a JWT token obtained from `/auth/login`. Include the token in the Authorization header:
    ```
    Authorization: Bearer {token}
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@savetogether.com
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Development Server
  - url: https://api.savetogether.com
    description: Production Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login endpoint. Format is "Bearer {token}"
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        image:
          type: string
          nullable: true
        bulk_price:
          type: number
          format: decimal
          description: Price per unit for bulk orders
        retail_price:
          type: number
          format: decimal
          description: Regular retail price
        moq:
          type: integer
          description: Minimum Order Quantity
          minimum: 1
        vendor_id:
          type: string
          format: uuid
        category:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, name, bulk_price, retail_price, moq, vendor_id]
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
          description: Unique order identifier
        group_id:
          type: string
          format: uuid
          nullable: true
        buyer_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, paid, confirmed, shipped, delivered, cancelled]
        total_amount:
          type: number
          format: decimal
        estimated_delivery:
          type: string
          format: date-time
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              product_id:
                type: string
                format: uuid
              product:
                $ref: '#/components/schemas/Product'
              quantity:
                type: integer
              price:
                type: number
                format: decimal
        created_at:
          type: string
          format: date-time
      required: [id, order_number, status, total_amount]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
      required: [success, error]
...
paths:
...
  /vendors/{id}/orders:
    get:
      tags: [Vendors]
      summary: Vendor's orders
      security:
        - bearerAuth: []
      description: Vendor only. Get orders for vendor's products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vendor orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'

    post:
      tags: [Orders]
      summary: Create order
      security:
        - bearerAuth: []
      description: Create new order with items. Validates MOQ per product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    properties:
                      product_id:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
                        minimum: 1
                group_id:
                  type: string
                  format: uuid
                  nullable: true
              required: [items]
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Validation error or MOQ violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Order'
  /orders/{id}/status:
    put:
      tags: [Orders]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, paid, confirmed, shipped, delivered, cancelled]
              required: [status]
      responses:
        '200':
          description: Order status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Order'

  # ============ ESCROW ============
...